ghostwriter/requirements.txt
openai>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
faiss-cpu>=1.7.0

# dev / test
pytest>=8.0
pytest-mock>=3.0
